+++ webpack.config.js
@@ -84,7 +84,6 @@ module.exports = (env, argv) => {
     output: {
       path: path.resolve(__dirname, "./dist"),
       filename: "./[name].js",
-      clean: true,
     },
     module: {
       rules: [