+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Un-Released]
 ### Enhancements
 - Firefox build scripts
+- Debug scripts
 
 
 ## [2.4.4]